+++ src/lib/st-dropdown-menu/st-dropdown-menu.interface.ts
@@ -15,9 +15,10 @@ export class StDropDownMenuItem {
    label: string;
    value: any;
    icon?: string;
+   labelColor?: string;
+   iconColor?: string;
    selected?: boolean;
    hasHtml?: boolean;
-
    [key: string]: any; // To do model more extensible if any other component needs to send more data
 }