+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * st-color-picker: Create component
 * st-two-list-selection: Allow to show dropdown menu options  if it is a non-editable list.
+* st-dropdown-menu: Allow to custom label and icon colors
 
 **Fixed bugs:**